requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.58.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
numpy==1.24.2
oauthlib==3.2.2
pandas==2.0.0
pycparser==2.21
PyJWT==2.8.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
views==0.3
Werkzeug==3.0.1
XlsxWriter==3.1.0
zipp==3.17.0
